# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.1
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine supports three auth modes derived from configuration:

    - **Open mode**: no `AUTH_TOKEN` and no `ADMIN_PASSWORD`; requests do not
      require authentication.
    - **Token mode**: `AUTH_TOKEN` set and `ADMIN_PASSWORD` unset; clients use
      `Authorization: Bearer <token>` or `?token=`.
    - **Full mode**: `ADMIN_PASSWORD` set; browser users authenticate with JWT
      sessions and service clients can use `tre_...` API keys. If `AUTH_TOKEN`
      is also set, it acts as a public read token returned by `/auth-init`.

    Mutating operations (POST, PATCH, PUT, DELETE) require editor/admin JWT
    role, a suitable API key, or the deprecated legacy `WRITE_TOKEN`.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
...

...
      specific site.

    Example:
    ```
...
    ```

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /unit-tag-suggestions:
    get:
      operationId: listUnitTagSuggestions
      summary: List unit alpha tag suggestions
      description: |
        Returns candidate unit alpha tags extracted from transcriptions where
        a unit identified itself (e.g. "County, Medic 12 on scene"). Only the
        unit's own speech is used (word-level attribution against the call's
        `src_list`); transcriptions without attribution are skipped unless
        every transmission in the call came from the same unit.

        Only unit designators are extracted ("Medic 12", "Battalion Chief 2",
        "P338", "1 Paul 31", "Adam 12"), never personal names, and only from a
        call's primary (current) transcription. A transmission recorded by
        several sites (one call group) counts as one call.

        Candidates are keyed by `(system_id, unit_id, tag_key)` and accumulate
        evidence across calls. **Pending** suggestions are only listed once they
        pass the review gate: seen in at least `UNIT_TAG_SUGGESTIONS_MIN_CALLS`
        distinct calls (default 3), at least `UNIT_TAG_SUGGESTIONS_MIN_SHARE`
        (default 0.2) of the unit's self-identification calls, and not already
        the unit's alpha tag. The gate applies to pending rows for every
        `status` value; approved and dismissed rows are always listed.

        Pending results are grouped by unit (busiest units first, then each
        unit's candidates by `call_count`); approved/dismissed results are
        ordered by decision time, newest first. The endpoint works whether or
        not the background scanner is enabled — `scanner.enabled` reports it.
      tags: [unit-tag-suggestions]
      parameters:
        - name: status
          in: query
          description: Filter by review status (`all` = every status).
          schema:
            type: string
            enum: [pending, approved, dismissed, all]
            default: pending
        - name: system_id
          in: query
          description: Filter by system ID (comma-separated for multiple). Alias `systems`.
          schema:
            type: string
            example: "1,2"
        - name: unit_id
          in: query
          description: Filter by radio unit ID (comma-separated for multiple). Alias `units`.
          schema:
            type: string
            example: "1234567"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitTagSuggestionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /unit-tag-suggestions/{id}:
    get:
      operationId: getUnitTagSuggestion
      summary: Get a unit alpha tag suggestion
      description: Returns one suggestion by ID regardless of status or review gate.
      tags: [unit-tag-suggestions]
      parameters:
        - $ref: "#/components/parameters/suggestionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitTagSuggestion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /unit-tag-suggestions/{id}/approve:
    post:
      operationId: approveUnitTagSuggestion
      summary: Approve a unit alpha tag suggestion
      description: |
        Applies a **pending** suggestion to its unit and marks it `approved`.
        The optional body overrides the proposed tag ("edit, then approve");
        without it the suggestion's `proposed_tag` is applied.

        The unit is updated through the same path as `PATCH /units/{id}`
        with `alpha_tag_source` set to `manual` (a person explicitly chose the
        tag, so MQTT and CSV re-imports will not overwrite it), including the
        best-effort CSV writeback when `CSV_WRITEBACK` is enabled. The unit's
        previous tag and source are recorded on the suggestion
        (`previous_tag`, `previous_tag_source`) with `decided_by`/`decided_at`.
        Other suggestions for the same unit are left untouched. Requires write
        access (same as `PATCH /units/{id}`).
      tags: [unit-tag-suggestions]
      parameters:
        - $ref: "#/components/parameters/suggestionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitTagSuggestionApprove"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitTagSuggestionApproveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Suggestion (or its unit) not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          $ref: "#/components/responses/SuggestionNotPending"
        "500":
          $ref: "#/components/responses/InternalError"
  /unit-tag-suggestions/{id}/dismiss:
    post:
      operationId: dismissUnitTagSuggestion
      summary: Dismiss a unit alpha tag suggestion
      description: |
        Marks a **pending** suggestion `dismissed` without touching the unit.
        A dismissed candidate keeps accumulating evidence when seen again but
        never returns to the pending queue. No request body. Requires write
        access (same as `PATCH /units/{id}`).
      tags: [unit-tag-suggestions]
      parameters:
        - $ref: "#/components/parameters/suggestionId"
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitTagSuggestionDismissResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/SuggestionNotPending"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Unit Events (system-wide)
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
...
    suggestionId:
      name: id
      in: path
      required: true
      description: Unit tag suggestion ID
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 42
...
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Unauthorized — missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden — valid read token but write token required for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SuggestionNotPending:
      description: |
        The suggestion has already been decided. `detail` carries the current
        status, e.g. `status: approved`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: conflict
            error: unit tag suggestion is approved, not pending
            detail: "status: approved"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    EventType:
      type: string
      enum: [on, off, join, call, end, data, ans_req, location, ackresp, signal, call_alert]
      description: |
        Unit event types from trunk-recorder MQTT:
        - **on**: unit registration (radio powered on / entering system)
        - **off**: unit deregistration (radio powered off / leaving system)
        - **join**: talkgroup affiliation (unit joins a talkgroup)
        - **call**: channel grant (voice call participation)
        - **end**: transmission end (per-unit end-of-transmission report)
        - **data**: data grant
        - **ans_req**: answer request
        - **location**: unit location update
        - **ackresp**: acknowledge response
        - **signal**: in-band signaling decode (MDC1200, FleetSync, STAR)
        - **call_alert**: directed unit-to-unit page (P25 TSBK 0x1F); `target_unit` and `target_unit_alpha_tag` present in SSE payload and `metadata_json`
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    Unit:
      type: object
      description: A radio unit (mobile or portable radio)
      required: [system_id, unit_id]
      properties:
        system_id:
          type: integer
          description: Database system ID this unit belongs to
          example: 1
        system_name:
          type: string
          description: System display name (embedded to avoid lookup)
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"
        unit_id:
          type: integer
          example: 1234567
        alpha_tag:
          type: string
          example: Engine 1
        alpha_tag_source:
          type: string
          description: |
            Where the alpha tag came from: `manual` (user edit), `csv` (unit tags
            CSV via TR_DIR or `POST /unit-tags/import`), or `mqtt`. Omitted when
            the tag was discovered from MQTT/audio metadata. Priority is
            manual > csv > mqtt; lower-priority sources never overwrite higher ones.
          example: csv
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        # Last event context (embedded)
        last_event_type:
          $ref: "#/components/schemas/EventType"
        last_event_time:
          type: string
          format: date-time
        last_event_tgid:
          type: integer
          example: 9178
        last_event_tg_tag:
          type: string
          description: Alpha tag of the talkgroup from last event
          example: "09-8L Main"
        # Tag observations (auto-managed, read-only). Stored beside alpha_tag
        # and never change it or its manual > csv > mqtt priority.
        recorder_alpha_tag:
          type: string
          readOnly: true
          description: |
            Latest non-empty unit tag reported by trunk-recorder (`unit_alpha_tag`
            in MQTT unit/call messages, or the srcList `tag` in watch/upload
            modes), recorded even when a manual or CSV `alpha_tag` wins. This is
            trunk-recorder's own resolved tag: its `unitTagsFile` entry if it has
            one, otherwise the decoded over-the-air alias. Omitted if never reported.
          example: "P338 FF1"
        recorder_alpha_tag_seen:
          type: string
          format: date-time
          readOnly: true
          description: When `recorder_alpha_tag` was last reported.
        ota_alpha_tag:
          type: string
          readOnly: true
          description: |
            Latest raw over-the-air unit alias, as reported separately from the
            resolved tag: `unit_alpha_tag_ota` in MQTT unit/call messages (MQTT
            plugin), or `srcList[].tag_ota` in trunk-recorder 5.2+ call JSON
            (watch/upload modes). Latest known value only (not a history);
            replaced when a different alias is reported. trunk-recorder reports
            the latest alias it knows for the unit, so the seen times record when
            tr-engine received that report, not necessarily a fresh decode.
            Omitted until a source that sends it reports this unit.
          example: "P338 FF1"
        ota_alpha_tag_first_seen:
          type: string
          format: date-time
          readOnly: true
          description: |
            When tr-engine first recorded the current `ota_alpha_tag` value.
            Resets when the alias changes. A late report of the same alias with
            an older timestamp does not move it earlier, so it never reaches back
            past an earlier alias change (archive import and system merge move it
            earlier only when the incoming observation window overlaps).
        ota_alpha_tag_last_seen:
          type: string
          format: date-time
          readOnly: true
          description: When the current `ota_alpha_tag` value was last observed.
        call_count:
          type: integer
          description: "Number of calls within the requested time window (only present on /talkgroups/{id}/units)"
          example: 42
        # Search field
        relevance_score:
          type: integer
          description: "Search relevance: 100=exact, 50=prefix, 10=contains. Only present in search results."
          example: 100
...
    UnitTagEvidence:
      type: object
      description: |
        One call in which the unit identified itself with the candidate tag.
        Suggestions keep the 10 most recent (newest first).
      required: [call_id, call_start_time, transcription_id, tgid, excerpt, pattern, start, end, audio_url]
      properties:
        call_id:
          type: integer
          format: int64
          example: 48531
        call_group_id:
          type: integer
          format: int64
          description: |
            Call group of the call (omitted when it has none). Other sites'
            recordings of the same transmission share it and are not counted again.
          example: 20417
        call_start_time:
          type: string
          format: date-time
        transcription_id:
          type: integer
          format: int64
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup alpha tag at the time of the call (omitted when empty)
          example: "BC Fire Dispatch"
        excerpt:
          type: string
          description: The unit's own transmission text (max 240 characters)
          example: "County, Medic 12 on scene."
        pattern:
          type: string
          description: |
            Extraction rule that matched: `addressee_caller` ("County, Medic 12 ..."),
            `unit_caller` ("Engine 5, Medic 12."), `caller_to` ("Medic 12 to County"),
            `from_caller` ("Command from Engine 5"), `this_is` ("this is Rescue 2"),
            `status` ("Engine 5 on scene"), `bare` (the whole transmission is the ID).
          enum: [addressee_caller, unit_caller, caller_to, from_caller, this_is, status, bare]
        start:
          type: number
          format: double
          description: Seconds into the call audio where the unit's transmission starts
        end:
          type: number
          format: double
          description: Seconds into the call audio where the unit's transmission ends
        audio_url:
          type: string
          description: Call audio URL (same as `GET /calls/{id}/audio`; 404 if the call has no audio)
          example: /api/v1/calls/48531/audio
    UnitTagSuggestion:
      type: object
      description: A candidate alpha tag for one unit, with accumulated evidence.
      required:
        - id
        - system_id
        - unit_id
        - unit_alpha_tag
        - tag_key
        - proposed_tag
        - status
        - occurrences
        - call_count
        - share
        - matches_current_tag
        - first_seen
        - last_seen
        - applied_tag
        - previous_tag
        - previous_tag_source
        - decided_at
        - decided_by
        - created_at
        - updated_at
        - evidence
      properties:
        id:
          type: integer
          format: int64
          example: 42
        system_id:
          type: integer
          example: 1
        system_name:
          type: string
          example: "Butler/Warren P25"
        unit_id:
          type: integer
          example: 1234567
        unit_alpha_tag:
          type: string
          description: The unit's current alpha tag ("" if none)
          example: ""
        unit_alpha_tag_source:
          type: string
          description: Source of the unit's current alpha tag (omitted when empty)
          example: mqtt
        tag_key:
          type: string
          description: Normalized comparison key (uppercase, spoken numbers as digits)
          example: "MEDIC 12"
        proposed_tag:
          type: string
          description: Display form of the candidate
          example: "Medic 12"
        status:
          type: string
          enum: [pending, approved, dismissed]
        occurrences:
          type: integer
          description: Total self-identifications with this tag (a call can contribute several)
          example: 6
        call_count:
          type: integer
          description: Distinct calls with this self-identification (multi-site recordings of one transmission count once)
          example: 4
        share:
          type: number
          format: double
          description: This candidate's call_count as a fraction of all of the unit's candidates' call_counts
          example: 0.8
        matches_current_tag:
          type: boolean
          description: |
            True when the unit's alpha tag already names this designator —
            the whole tag, a designator inside it ("BCFD Medic 12"), or its
            phonetic letter form ("1P31" for "1 Paul 31"). Computed at the last
            sighting; if the unit's tag has changed since, it is re-checked
            live against the new tag (whole tag or whole words inside it,
            ignoring case and punctuation; phonetic forms are re-checked at
            the next sighting). Such pending rows are not listed.
        first_seen:
          type: string
          format: date-time
          description: Start time of the earliest call with this self-identification
        last_seen:
          type: string
          format: date-time
        applied_tag:
          type: string
          nullable: true
          description: Tag written to the unit on approval (proposed_tag or the override)
        previous_tag:
          type: string
          nullable: true
          description: The unit's alpha tag before approval
        previous_tag_source:
          type: string
          nullable: true
          description: The unit's alpha_tag_source before approval
        decided_at:
          type: string
          format: date-time
          nullable: true
        decided_by:
          type: string
          nullable: true
          description: Username or API key label of the reviewer (null in open/token mode)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/UnitTagEvidence"
    UnitTagScannerStatus:
      type: object
      description: Background scanner configuration and backfill progress.
      required: [enabled, min_calls, min_share, last_transcription_id, max_transcription_id, updated_at]
      properties:
        enabled:
          type: boolean
          description: Whether the scanner runs in this process (`UNIT_TAG_SUGGESTIONS`)
        min_calls:
          type: integer
          description: Review gate — `UNIT_TAG_SUGGESTIONS_MIN_CALLS`
          example: 3
        min_share:
          type: number
          format: double
          description: Review gate — `UNIT_TAG_SUGGESTIONS_MIN_SHARE`
          example: 0.2
        last_transcription_id:
          type: integer
          format: int64
          description: Highest transcription ID processed (0 if never run)
        max_transcription_id:
          type: integer
          format: int64
          description: Highest transcription ID in the database; the scanner is caught up when equal
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: Last time the scanner advanced (null if it never ran)
    UnitTagSuggestionApprove:
      type: object
      description: Optional approval body. Omit (or send `{}`) to apply `proposed_tag`.
      properties:
        alpha_tag:
          type: string
          description: Tag to apply instead of `proposed_tag` (trimmed; must not be blank)
          example: "BCFD Medic 12"

    # ========== Response Wrappers ==========
...
    UnitTagSuggestionListResponse:
      type: object
      required: [suggestions, total, limit, offset, scanner]
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/UnitTagSuggestion"
        total:
          type: integer
          example: 12
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
        scanner:
          $ref: "#/components/schemas/UnitTagScannerStatus"
    UnitTagSuggestionApproveResponse:
      type: object
      required: [suggestion, unit]
      properties:
        suggestion:
          $ref: "#/components/schemas/UnitTagSuggestion"
        unit:
          $ref: "#/components/schemas/Unit"
    UnitTagSuggestionDismissResponse:
      type: object
      required: [suggestion]
      properties:
        suggestion:
          $ref: "#/components/schemas/UnitTagSuggestion"
...